+++ src/game/FileManager.h
@@ -34,11 +34,11 @@ private:
 	CStrList SearchDirHead;
 	CStrList PackedFileHead;
 	CStr FileName;
-	unsigned long FileOffset;
+	uint32_t FileOffset;
 	//
 	FILE *FileStack[FILE_STACK_SIZE];
 	CStr FileNameStack[FILE_STACK_SIZE];
-	unsigned long FileOffsetStack[FILE_STACK_SIZE];
+	uint32_t FileOffsetStack[FILE_STACK_SIZE];
 	int nFileStack;
 public:
 	//Access members.
@@ -58,7 +58,7 @@ public:
 	bool PopFile();	//Returns true if file stack was successfully popped, false if stack is at bottom.
 	//
 	CStr GetFileName();
-	unsigned long GetFileOffset();	//These are for external routines that MUST open a file via name and offset.
+	uint32_t GetFileOffset();	//These are for external routines that MUST open a file via name and offset.
 	void Close();
 	//These work on the currently opened file, either physical or in pack.
 	size_t fread(void *buf, size_t size, size_t count);
